fix: SSM recurrent-state integrity clamp + 256K-context forward verification
Root cause of decode-after-prefill NaN: the Gated DeltaNet recurrent
state (model->ssm_states, persistent across calls) had no upper bound.
For untrained/random SSM weights (and any real model under a transient
positive-mean gate spike) the per-chunk decay exp(g_last) > 1 pumps the
state to Inf/NaN within a few chunks, permanently poisoning the persistent
state so every later forward (decode after a prefill) was NaN.

Fix: add ssm_state_clamp() (threshold 1e3f) applied after every state
update in both recurrence paths (wubu_ssm_chunked.c chunked; wubu_ssm.c
scalar + save path). Trained models keep state well below threshold
(no-op); divergent weights get a hard floor that keeps the model
re-callable.

Also includes prior 256K-path fixes verified by ASAN:
- wubu_ssm.c:249 int-overflow guard on B*T*C*k
- wubu_model.c:947/977 int-overflow guard on N*vocab_size
- wubu_moe.c: alloca->heap for gate_shared/up_shared (stack-overflow fix)

Verification:
- test_256k_forward: prefill finite at T=126000 + decode-after-prefill
  finite (same-model T=1) + fresh-model decode finite. ALL CHECK PASSED.
- divergence probe: ssm_nan=0 at all T through 126000.
- regression suite green: test_st_bridge, test_btl3_lora, test_new_models.

Add docs/AGENTIC_256K_PLAN.md (slermes agentic harness plan).

+++ src/wubu_model.c
@@ -944,7 +944,7 @@ void wubu_model_forward_from_embd(wubu_model_t *model,
         // Compare against F32 SGEMM when output_weight is also loaded
         if (model->output_weight && getenv("VERBOSE_OUTPUT_PROJ")) {
             float *f32_logits = (float *)malloc(N * model->vocab_size * sizeof(float));
-            #pragma omp parallel for collapse(2) if(N * model->vocab_size > 100000)
+            #pragma omp parallel for collapse(2) if((int64_t)N * model->vocab_size > 100000)
             for (int i = 0; i < N; i++) {
                 for (int j = 0; j < model->vocab_size; j++) {
                     const float *h_i = x + i * model->d_model;
@@ -974,7 +974,7 @@ void wubu_model_forward_from_embd(wubu_model_t *model,
     } else if (model->output_weight) {
         // F32 output projection: logits[v] = sum_k x[k] * output_weight[v*d_model + k]
         // (F32 safetensors/HF path: output_weight_q is NULL, output_weight holds plain f32 lm_head)
-        #pragma omp parallel for if(N * model->vocab_size > 100000)
+        #pragma omp parallel for if((int64_t)N * model->vocab_size > 100000)
         for (int i = 0; i < N; i++) {
             const float *h_i = x + i * model->d_model;
             float *log_i = logits + i * model->vocab_size;
